% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.lm.R
\name{my.lm}
\alias{my.lm}
\title{Linear Regression Model}
\usage{
my.lm(x, y)
}
\arguments{
\item{x}{a dataframe contains all observations in selected columns from a dataset. These columns
are the covariates that you are interested in.}

\item{y}{The target variable to fit the linear regression model on.}
}
\value{


\item{R-squared value}{A value which is calculated by SSR/SSY. It can be interpreted as the
percentage of variance that can be explained by the model. Higher R-squared value means the model
fits better.}

\item{F statistic}{This value shows if there is any of the selected variables in your input
x is related to y. This value is calculated by (R^2/1-R^2)*(df2/df1)}

\item{df2}{The degree of freedom of error, which is calculated by n-p}

\item{df1}{The degree of freedom of regression, which is calculated by p-1}

\item{Adjusted R-squared value}{When we have too many covariates, the R-squared value will
be penalized, this new value is called adjusted R-squared value}

\item{Residuals squared vs. fitted values plot}{This plot is used to check if
the residuals are randomly distributed around 0.}

\item{Normal Q-Q Plot}{This plot is used to check the assumption of normality}

\item{Residuals squared vs. fitted values plot}{This plot is used to check if
the residuals squared are randomly distributed around 0.}
}
\description{
This my.lm function is used to fit linear models. It can be used to get the coefficients
from a linear regression model. In addition this function will also output R squared value,
adjusted R-squared value, F statistic with degree of freedom, and three plots to perform residuals
diagnostic to check the assumptions of linear regression model.
}
\examples{
## A linear regression model to test the effect of disp,
hp and drat on mpg.
x = mtcars[, c(3,4,5)]
y = mtcars$mpg
mymodel <- my.lm(x, y)

}
